server/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.1.6
django-admin-interface==0.20.0
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-colorfield==0.7.3
django-cors-headers==3.13.0
django-enum-choices==2.1.4
django-environ==0.9.0
django-filter==22.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-redis==5.2.0
django-storages==1.13.1
django-timezone-field==5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
flower==1.2.0
gunicorn==20.1.0
humanize==4.6.0
kombu==5.2.4
packaging==23.0
Pillow==8.3.2
prometheus-client==0.16.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.1
PyJWT==2.6.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2022.7.1
redis==4.2.0
six==1.16.0
sqlparse==0.4.3
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.7
vine==5.0.0
wcwidth==0.2.6
wrapt==1.14.1
